Require at least Rust 1.4.0 for examples
I use the `result.expect()` method, which was not stable until
Rust 1.4.0. The crate still compiles with all stable versions of Rust.
For CI, I will not test those older versions. It would be possible to
only build and not test the older versions of Rust, but the extra
complexity is not worth it.