Deal with libc incompatibilities on Mac OS
The libc crate on Unix-ish platforms (excluding Mac OS) has a mincore
function that takes an unsigned char. However, on Mac OS, it takes a
signed char, leading to compilation errors because of the type mismatch.
The data is the same anyway, so it is only the type that needs to be
adapted.

To fix this issue, just copy the entire function into a Mac OS-specific
module, with an updated type. Perhaps not the cleanest solution, but I
don't expect to have to change this code a lot, so it is fine.