Work around breaking change in time crate
The time crate released a new version, v0.1.36, which introduces
breaking changes for Rust 1.4 through Rust 1.7. However, the rust-crypto
crate depends on "time ^0.1", so Cargo will select the broken version.
Work around this by locking the version to the previous one. As this is
only a dev-dependency, this does not affect users of Filebuffer.

Breaking changes were introduced in:

  https://github.com/rust-lang-deprecated/time/pull/140
  https://github.com/rust-lang-deprecated/time/pull/142