Put Cargo.lock under source control
Yes, this is a library. But this is the second time now, that a
dependency had a breaking change without bumping semver (to be fair,
these crates are still unstable, but de facto they have been stable for
years). I want my crates to keep working on CI regardless of external
changes, so that means putting Cargo.lock under source control.