Bump libc dependency to 2.86
Also drop support for older Rust versions, make 1.25 the new minimum.
Reflect this in the CI configs too. The new libc version is required to
avoid linker errors on Apple M1 systems.

This is an alternative to pull request 8 by Chad Brewbaker, which also
removes some now-outdated overrides, but avoids reformatting the code.
Thanks Chad for reporting this!